requirements.txt
feedparser
PyYAML

